Process a Plate and Save Normalised Output to CSV

This function processes a plate and computes normalised values based on the specified normalisation_type. The function supports three types of normalisation:

RAU (Relative Antibody Units) - Default normalisation type.

nMFI (Normalised Median Fluorescence Intensity).

MFI (Median Fluorescence Intensity, blank-adjusted raw MFI values).

Depending on the chosen normalisation type, the function:

Adjusts for blanks (if blank_adjustment = TRUE).

Computes the relevant normalised values for each analyte in the plate.

Aggregates computed values into a single data frame.

Optionally writes the results to a CSV file.

RAU Normalisation Workflow:

Blank adjustment (if enabled).

Fit standard curve models for each analyte.

Compute RAU values based on the fitted models.

Aggregate RAU values into a data frame.

Save the results to a CSV file.

nMFI Normalisation Workflow:

Blank adjustment (if enabled).

Compute nMFI values using the target dilution.

Aggregate nMFI values into a data frame.

Save the results to a CSV file.

MFI Normalisation Workflow:

Blank adjustment (if enabled).

Save the adjusted MFI values to a CSV file.

If the plate is already blank-adjusted, setting blank_adjustment = TRUE has no effect.

More details on the normalisation methods can be found in:

RAU: create_standard_curve_model_analyte.

nMFI: get_nmfi.

Usage

process_plate(
  plate,
  filename = NULL,
  output_dir = "normalised_data",
  write_output = TRUE,
  normalisation_type = "RAU",
  data_type = "Median",
  blank_adjustment = FALSE,
  verbose = TRUE,
  reference_dilution = 1/400,
  ...
)

Arguments

plate

(Plate) A plate object containing fluorescence intensity data.

filename

(character(1), optional) Output CSV filename. Defaults to {plate_name}_{normalisation_type}.csv.

If omitted (NULL), the filename is auto-generated.

If the filename lacks .csv, the extension is automatically added.

If output_dir is specified, it is combined with filename unless filename is an absolute path.

If a file already exists at the path, it will be overwritten.

output_dir

(character(1), default = 'normalised_data') Directory for saving the output file.

If the directory does not exist, it will be created.

If NULL, the current working directory is used.

write_output

(logical(1), default = TRUE) Whether to save the output to a CSV file.

normalisation_type

(character(1), default = 'RAU') The normalisation method to apply.

Allowed values: c(MFI, RAU, nMFI).

data_type

(character(1), default = 'Median') The data type to use for calculations.

blank_adjustment

(logical(1), default = FALSE) Whether to perform blank adjustment before computing values.

verbose

(logical(1), default = TRUE) Whether to print additional information during execution.

reference_dilution

(numeric(1) or character(1), default = 1/400)

The target dilution used as a reference for nMFI normalisation.

Ignored for RAU normalisation.

Can be numeric (e.g., 0.0025) or a string ('1/400').

...

Additional arguments passed to the model fitting function (create_standard_curve_model_analyte).

Value

A data frame containing the computed normalised values.

Examples

plate_file <- system.file("extdata", "CovidOISExPONTENT_CO_reduced.csv", package = "SerolyzeR")
layout_file <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzeR")
plate <- read_luminex_data(plate_file, layout_file, verbose = FALSE)
#> Could not parse datetime string using default datetime format. Trying other possibilies.
#> Successfully parsed datetime string using order: mdY IM p
example_dir <- tempdir(check = TRUE)

# Process plate with default settings (RAU normalisation)
process_plate(plate, output_dir = example_dir)
#> Warning: The specified file /tmp/Rtmpn5LAtP/CovidOISExPONTENT_CO_reduced_RAU.csv already exists. Overwriting it.
#> Fitting the models and predicting RAU for each analyte
#> Saving the computed RAU values to a CSV file located in: '/tmp/Rtmpn5LAtP/CovidOISExPONTENT_CO_reduced_RAU.csv'
#>         Spike_6P_IPP      ME_NA
#> K086-LM   3292.26101  2128.6098
#> M254-VM     85.38753  5496.4318
#> M199-DS   2841.64505  7169.3408
#> M164-LM    125.42694  3560.9584
#> M265-MM    396.34769  3783.9004
#> K026-DJ     85.13506  1430.5461
#> K137-DT   3142.28223  6567.6666
#> M255-VA     38.78304  1103.7202
#> M258-PA     52.98592  2174.6786
#> M188-VC   1137.87727  1877.6962
#> M270-BF     58.80200   875.6587
#> M050-EL    100.64396  2168.0885
#> M088-GE    145.30638  2892.0240
#> M259-PM   1082.02893  8614.0967
#> K101-PA     73.84503  1777.7663
#> M189-VY    511.61479  2562.3316
#> K018-FC    882.55266  2250.6828
#> M142-RA     54.43368   966.1990
#> K100-CC     42.98563   984.9380
#> K107-RP     32.36841  1138.1493
#> M240-HS     31.46594  1028.6766
#> K019-FM     83.62164  1589.0900
#> M148-PS     61.97864  1216.4977
#> M089-HG   4677.61877  7712.7546
#> K136-DA  20000.00000 17238.7897
#> M241-HR   5532.92911  1279.2952
#> K020-NA  12680.55047  9631.3814
#> M239-HJ     40.41069  2654.1971
#> M092-LS    164.84476  3102.6112
#> M162-PE   4972.48078  5738.5101
#> M260-PM  20000.00000  2370.4746
#> K021-FS     28.78211   635.1388
#> M253-VM     80.35078  1235.3249
#> M198-CN     60.26598  2056.5039
#> M163-PD   3780.09488  6970.3730
#> M264-MA  20000.00000  2197.7678
#> K024-DT    981.88707  2867.6895

# Process plate without blank adjustment, custom filename
process_plate(plate,
  filename = "plate_no_blank_adjustment.csv",
  output_dir = example_dir, blank_adjustment = FALSE
)
#> Fitting the models and predicting RAU for each analyte
#> Saving the computed RAU values to a CSV file located in: '/tmp/Rtmpn5LAtP/plate_no_blank_adjustment.csv'
#>         Spike_6P_IPP      ME_NA
#> K086-LM   3292.26101  2128.6098
#> M254-VM     85.38753  5496.4318
#> M199-DS   2841.64505  7169.3408
#> M164-LM    125.42694  3560.9584
#> M265-MM    396.34769  3783.9004
#> K026-DJ     85.13506  1430.5461
#> K137-DT   3142.28223  6567.6666
#> M255-VA     38.78304  1103.7202
#> M258-PA     52.98592  2174.6786
#> M188-VC   1137.87727  1877.6962
#> M270-BF     58.80200   875.6587
#> M050-EL    100.64396  2168.0885
#> M088-GE    145.30638  2892.0240
#> M259-PM   1082.02893  8614.0967
#> K101-PA     73.84503  1777.7663
#> M189-VY    511.61479  2562.3316
#> K018-FC    882.55266  2250.6828
#> M142-RA     54.43368   966.1990
#> K100-CC     42.98563   984.9380
#> K107-RP     32.36841  1138.1493
#> M240-HS     31.46594  1028.6766
#> K019-FM     83.62164  1589.0900
#> M148-PS     61.97864  1216.4977
#> M089-HG   4677.61877  7712.7546
#> K136-DA  20000.00000 17238.7897
#> M241-HR   5532.92911  1279.2952
#> K020-NA  12680.55047  9631.3814
#> M239-HJ     40.41069  2654.1971
#> M092-LS    164.84476  3102.6112
#> M162-PE   4972.48078  5738.5101
#> M260-PM  20000.00000  2370.4746
#> K021-FS     28.78211   635.1388
#> M253-VM     80.35078  1235.3249
#> M198-CN     60.26598  2056.5039
#> M163-PD   3780.09488  6970.3730
#> M264-MA  20000.00000  2197.7678
#> K024-DT    981.88707  2867.6895

# Process plate with nMFI normalisation
process_plate(plate,
  output_dir = example_dir, normalisation_type = "nMFI",
  reference_dilution = 1 / 400
)
#> Warning: The specified file /tmp/Rtmpn5LAtP/CovidOISExPONTENT_CO_reduced_nMFI.csv already exists. Overwriting it.
#> Computing nMFI values for each analyte
#> Saving the computed nMFI values to a CSV file located in: '/tmp/Rtmpn5LAtP/CovidOISExPONTENT_CO_reduced_nMFI.csv'
#>         Spike_6P_IPP     ME_NA
#> K086-LM   1.32720798 0.9407979
#> M254-VM   0.05968661 2.0952381
#> M199-DS   1.19088319 2.5559846
#> M164-LM   0.08190883 1.4723295
#> M265-MM   0.22336182 1.5495495
#> K026-DJ   0.05954416 0.6615187
#> K137-DT   1.28290598 2.3976834
#> M255-VA   0.03247863 0.5276705
#> M258-PA   0.04102564 0.9588160
#> M188-VC   0.56766382 0.8416988
#> M270-BF   0.04444444 0.4337194
#> M050-EL   0.06823362 0.9562420
#> M088-GE   0.09273504 1.2316602
#> M259-PM   0.54358974 2.9060489
#> K101-PA   0.05313390 0.8018018
#> M189-VY   0.28062678 1.1081081
#> K018-FC   0.45527066 0.9884170
#> M142-RA   0.04188034 0.4710425
#> K100-CC   0.03504274 0.4787645
#> K107-RP   0.02849003 0.5418275
#> M240-HS   0.02792023 0.4967825
#> K019-FM   0.05868946 0.7258687
#> M148-PS   0.04629630 0.5740026
#> M089-HG   1.69344729 2.6924067
#> K136-DA   4.10655271 4.3912484
#> M241-HR   1.88774929 0.5997426
#> K020-NA   2.98148148 3.1299871
#> M239-HJ   0.03347578 1.1428571
#> M092-LS   0.10327635 1.3088803
#> M162-PE   1.76282051 2.1660232
#> M260-PM   4.17592593 1.0347490
#> K021-FS   0.02621083 0.3346203
#> M253-VM   0.05683761 0.5817246
#> M198-CN   0.04529915 0.9124839
#> M163-PD   1.46452991 2.5045045
#> M264-MA   5.67492877 0.9678250
#> K024-DT   0.49971510 1.2226512